ongoing

+++ Crypto/Wist.Crypto/ProofOfWork/NoneProofOfWorkCalculation.cs
@@ -9,7 +9,7 @@ namespace Wist.Crypto.ProofOfWork
     {
         public POWType POWType => POWType.None;
 
-        public ushort HashSize => 0;
+        public int HashSize => 0;
 
         public byte[] CalculateHash(byte[] input)
         {
